+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Retrieve enrollment status for ongoing course runs
 - Add a shortcut route to redirect to the course page through a course code.
 - Add cunningham and design system.
 - Specific sidebar for order related routes